requirements.txt
weaviate-client==3.15.6
datasets==2.11.0
apache-beam==2.46.0